requirements.txt
psycopg2
googletran
python-dotenv